# server/swagger.yaml
openapi: 3.0.0
info:
  title: chytraskola
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  /user/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - User
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - User
      parameters:
        - name: id
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    post:
      summary: CRUD ACTION update
      tags:
        - User
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
    
  #-----  dochadzka  -----
...
  /dochadzka/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - dochadzka
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - dochadzka
      parameters:
        - name: id
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/dochadzka"
    post:
      summary: CRUD ACTION update
      tags:
        - dochadzka
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/dochadzka"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/dochadzka"
  /dochadzka/findBy_triedy/{key}:
    get:
      summary: CRUD ACTION findBy_triedy
      tags:
        - dochadzka
      parameters:
        - name: key
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/dochadzka"
  /dochadzka/findBy_ucitel/{key}:
    get:
      summary: CRUD ACTION findBy_ucitel
      tags:
        - dochadzka
      parameters:
        - name: key
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/dochadzka"
    
  #-----  odbory  -----
...
  /odbory/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - odbory
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - odbory
      parameters:
        - name: id
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/odbory"
    post:
      summary: CRUD ACTION update
      tags:
        - odbory
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/odbory"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/odbory"
    
  #-----  predmety  -----
...
  /skola/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - skola
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - skola
      parameters:
        - name: id
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/skola"
    post:
      summary: CRUD ACTION update
      tags:
        - skola
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/skola"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/skola"
    
  #-----  skupina  -----
...
  /Users/{id}/changePassword:
    post:
      summary: Change password by admin API
      tags:
        - User
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                passwordNew:
                  type: string
                passwordAdmin:
                  type: string
      responses:
        '200':
          description: OK
  /Users/{id}:
    post:
      summary: Update User API
      tags:
        - User
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
    delete:
      summary: Delete User API
      tags:
        - User
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
security:
  - bearerAuth: [] 
  
  
# Components
components:
...
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes

  # Data Model
  schemas:
...
    User:
      properties:
        _id:
          type: string
        mail:
          type: string
        name:
          type: string
        password:
          type: string
        roles:
          type: string
        surname:
          type: string
        username:
          type: string
      required:
        -  _id
        - password
        - username
    dochadzka:
      properties:
        _id:
          type: string
        datum:
          type: integer
        predmet:
          type: string
        status:
          type: string
        student:
          type: string
        trieda:
          type: string
        ucitel:
          type: string
      required:
        -  _id
    odbory:
      properties:
        _id:
          type: string
        forma_studia:
          type: string
        kod:
          type: string
        nazov:
          type: string
        orientacia:
          type: string
        popis:
          type: string
        ucebny_plan:
          type: string
      required:
        -  _id
...
    skola:
      properties:
        _id:
          type: string
        adresa:
          type: string
        datum_zalozenia:
          type: integer
        dic:
          type: string
        email:
          type: string
        facebook:
          type: string
        ico:
          type: string
        instagram:
          type: string
        logo:
          type: object
        mobil:
          type: string
        nazov:
          type: string
        riaditel:
          type: string
        veduca_sj:
          type: string
        web:
          type: string
        zastupca_riaditela:
          type: string
      required:
        -  _id
...